Revert "use <livewire > tag and move graph to upper level"
