+++ app/Http/Livewire/CovidState/Graph/ActiveCasePercentage.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Livewire\Graph;
+namespace App\Http\Livewire\CovidState\Graph;
 
 use Illuminate\Contracts\Foundation\Application;
 use Illuminate\Contracts\View\Factory;
@@ -8,15 +8,15 @@ use Illuminate\Contracts\View\View;
 use Illuminate\Support\Collection;
 use Livewire\Component;
 
-class CasePerDay extends Component
+class ActiveCasePercentage extends Component
 {
-
     public Collection $date;
-    public Collection $confirmCase;
-    public Collection $recoveredCase;
+    public Collection $activeCase;
+    public Collection $cumRecoveredCase;
+    public Collection $cumDeathCase;
 
     public function render(): Factory|View|Application
     {
-        return view('livewire.graph.case-per-day');
+        return view('livewire.covid-state.graph.active-case-percentage');
     }
 }
