+++ resources/views/livewire/covid-state/graph/active-case-percentage.blade.php
