+++ resources/views/livewire/covid-state/graph/active-case-vs-health-care.blade.php
