+++ resources/views/livewire/covid-state/graph/death-per-day.blade.php
